# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {},
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "SuccessEnvelope": {
        "properties": {
          "data": {
            "description": "仅包含沪深市场数据"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "证券代码使用 000001.SZ 或 600519.SH 格式；全市场接口使用 market=0（深圳）或 market=1（上海）。",
    "title": "沪深股票行情 API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/mac/boards/members": {
      "get": {
        "operationId": "mac.board.members",
        "parameters": [
          {
            "in": "query",
            "name": "board_symbol",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._-]{1,32}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询 MAC 板块成分"
      }
    },
...
    "/api/v1/mac/monitor": {
      "get": {
        "operationId": "mac.monitor",
        "parameters": [
          {
            "in": "query",
            "name": "market",
            "required": true,
            "schema": {
              "description": "0=深圳，1=上海",
              "enum": [
                0,
                1
              ],
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询 MAC 市场监控"
      }
    },
    "/api/v1/mac/symbols/auction": {
      "get": {
        "operationId": "mac.auction",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询 MAC 集合竞价"
      }
    },
...
    "/api/v1/mac/symbols/capital-flow": {
      "get": {
        "operationId": "mac.capital-flow",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询 MAC 资金流向"
      }
    },
...
    "/api/v1/mac/symbols/transactions": {
      "get": {
        "operationId": "mac.transactions",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "description": "YYYYMMDD；部分实时接口可省略",
              "maximum": 21001231,
              "minimum": 19900101,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询 MAC 成交"
      }
    },
...
    "/api/v1/stocks/auction": {
      "get": {
        "operationId": "stock.auction",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询集合竞价"
      }
    },
...
    "/api/v1/stocks/finance": {
      "get": {
        "operationId": "company.finance",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询财务信息"
      }
    },
...
    "/api/v1/stocks/index-momentum": {
      "get": {
        "operationId": "stock.index.momentum",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询指数动量"
      }
    },
    "/api/v1/stocks/orders/history": {
      "get": {
        "operationId": "stock.orders.history",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "description": "YYYYMMDD；部分实时接口可省略",
              "maximum": 21001231,
              "minimum": 19900101,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询历史委托"
      }
    },
...
    "/api/v1/stocks/sampling": {
      "get": {
        "operationId": "stock.sampling",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询分时采样"
      }
    },
...
    "/api/v1/stocks/ticks/history": {
      "get": {
        "operationId": "stock.tick.history",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "description": "YYYYMMDD；部分实时接口可省略",
              "maximum": 21001231,
              "minimum": 19900101,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询历史分时"
      }
    },
    "/api/v1/stocks/transactions": {
      "get": {
        "operationId": "stock.transactions",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询当日成交"
      }
    },
    "/api/v1/stocks/transactions/history": {
      "get": {
        "operationId": "stock.transactions.history",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "description": "YYYYMMDD；部分实时接口可省略",
              "maximum": 21001231,
              "minimum": 19900101,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询历史成交"
      }
    },
    "/api/v1/stocks/unusual": {
      "get": {
        "operationId": "stock.unusual",
        "parameters": [
          {
            "in": "query",
            "name": "market",
            "required": true,
            "schema": {
              "description": "0=深圳，1=上海",
              "enum": [
                0,
                1
              ],
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询市场异动"
      }
    },
    "/api/v1/stocks/volume-profile": {
      "get": {
        "operationId": "stock.volume-profile",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "examples": [
                "000001.SZ",
                "600519.SH"
              ],
              "pattern": "^[0-9]{6}\\.(SH|SZ)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "请求成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "参数错误"
          },
          "502": {
            "description": "上游不可用"
          },
          "504": {
            "description": "上游超时"
          }
        },
        "summary": "查询筹码分布"
      }
    },
...
  }
}